directory fix
