--- README.md
@@ -1,14 +1,0 @@
-# Pomodoro Timer ThirtyOne
-
-**Description:** 
-Pomodoro Timer project for CS 361 - Software Engineering
-
-**Group 31 Members:** 
-- Sarah Leon
-- Ryan Ong
-- Shawn Shahzad
-- Dingguo Tang
-- Ziwen Tong
-
-**User Interface Design:**
-https://app.lucidchart.com/invitations/accept/aca00859-9b66-4127-a0f7-9bee7a8f95ce
