--- combined/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Jul 25 11:39:59 PDT 2020
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
